codes/requirements.txt
numpy
opencv-python
pyyaml
tb-nightly
future
scp
tqdm
matplotlib
scipy
munch
tqdm
scp
tensorboard
pytorch_fid==0.1.1
kornia
linear_attention_transformer
vector_quantize_pytorch
orjson
einops